+++ dynomite-derive/src/lib.rs
@@ -39,7 +39,7 @@ use proc_macro_error::{abort, ResultExt};
 use quote::{quote, ToTokens};
 use syn::{
     parse::Parse, punctuated::Punctuated, Attribute, DataStruct, DeriveInput, Field, Fields, Ident,
-    Token, Visibility,
+    Path, Token, Visibility,
 };
 
 struct Variant {
@@ -71,7 +71,7 @@ impl DataEnum {
         attrs: &[Attribute],
     ) -> Self {
         let me = Self {
-            attrs: parse_attrs(&attrs),
+            attrs: parse_attrs(attrs),
             ident,
             variants: inner
                 .variants
@@ -118,9 +118,9 @@ impl DataEnum {
         let match_arms = self.variants.iter().map(|variant| {
             let variant_ident = &variant.inner.ident;
             let expr = match &variant.inner.fields {
-                Fields::Named(_record) => Self::unimplemented_record_variants(&variant),
+                Fields::Named(_record) => Self::unimplemented_record_variants(variant),
                 Fields::Unnamed(tuple) => {
-                    Self::expect_single_item_tuple(&tuple, variant_ident);
+                    Self::expect_single_item_tuple(tuple, variant_ident);
                     quote! { Self::#variant_ident(::dynomite::FromAttributes::from_attrs(attrs)?) }
                 }
                 Fields::Unit => quote! { Self::#variant_ident },
@@ -159,9 +159,9 @@ impl DataEnum {
             let variant_ident = &variant.inner.ident;
             let variant_deser_name = variant.deser_name();
             match &variant.inner.fields {
-                Fields::Named(_record) => Self::unimplemented_record_variants(&variant),
+                Fields::Named(_record) => Self::unimplemented_record_variants(variant),
                 Fields::Unnamed(tuple) => {
-                    Self::expect_single_item_tuple(&tuple, variant_ident);
+                    Self::expect_single_item_tuple(tuple, variant_ident);
 
                     quote! {
                         Self::#variant_ident(variant) => {
@@ -254,6 +254,13 @@ impl<'a> ItemField<'a> {
         self.attrs
             .iter()
             .any(|attr| matches!(attr.kind, FieldAttrKind::Default))
+    }
+
+    fn skip_serializing_if(&self) -> Option<&Path> {
+        self.attrs.iter().find_map(|attr| match &attr.kind {
+            FieldAttrKind::SkipSerializingIf(expr) => Some(expr),
+            _ => None,
+        })
     }
 
     fn is_flatten(&self) -> bool {
@@ -472,7 +479,7 @@ fn make_dynomite_attrs_for_struct(
     let from_attribute_map = get_from_attributes_trait(name, &item_fields);
     // impl ::dynomite::IntoAttributes for Name
     // impl From<Name> for ::dynomite::Attributes
-    let to_attribute_map = get_to_attribute_map_trait(name, &item_fields);
+    let to_attribute_map = get_into_attribute_map_trait(name, &item_fields);
     // impl TryFrom<::dynomite::Attributes> for Name
     // impl From<Name> for ::dynomite::Attributes
     let std_into_attrs = get_std_convert_traits(name);
@@ -506,7 +513,7 @@ fn make_dynomite_item(
     // impl ::dynomite::FromAttributes for Name
     let from_attribute_map = get_from_attributes_trait(name, &item_fields);
     // impl ::dynomite::IntoAttributes for Name
-    let to_attribute_map = get_to_attribute_map_trait(name, &item_fields);
+    let to_attribute_map = get_into_attribute_map_trait(name, &item_fields);
     // impl TryFrom<::dynomite::Attributes> for Name
     // impl From<Name> for ::dynomite::Attributes
     let std_into_attrs = get_std_convert_traits(name);
@@ -519,7 +526,7 @@ fn make_dynomite_item(
     })
 }
 
-fn get_to_attribute_map_trait(
+fn get_into_attribute_map_trait(
     name: &Ident,
     fields: &[ItemField],
 ) -> impl ToTokens {
@@ -557,17 +564,25 @@ fn get_into_attrs(fields: &[ItemField]) -> impl ToTokens {
         let field_deser_name = field.deser_name();
         let field_ident = &field.field.ident;
 
-        if field.is_flatten() {
+        let insert_attr = quote! {
+            attrs.insert(
+                #field_deser_name.to_string(),
+                ::dynomite::Attribute::into_attr(self.#field_ident)
+            );
+        };
+
+        if let Some(skip_serializing_if) = field.skip_serializing_if() {
+            quote! {
+                if !#skip_serializing_if(&self.#field_ident) {
+                    #insert_attr
+                }
+            }
+        } else if field.is_flatten() {
             quote! {
                 ::dynomite::IntoAttributes::into_attrs(self.#field_ident, attrs);
             }
         } else {
-            quote! {
-                attrs.insert(
-                    #field_deser_name.to_string(),
-                    ::dynomite::Attribute::into_attr(self.#field_ident)
-                );
-            }
+            insert_attr
         }
     });