+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.11.0
+
+* Introduce new `#[dynomite(skip_serializing_if = "...")]` field attribute that
+  allows for skipping the value from serializing into a map according to the
+  given condition. [#4](https://github.com/rust-serverless/dynomite/pull/4)
+* Fix manual impl of `Into` trait to `From` for retry policy [#4](https://github.com/rust-serverless/dynomite/pull/4)
+
 # 0.10.0
 
 * Bump rusoto dependencies to version from `0.44` to `0.45`